refactor: export `define_enum` macro
Also create `Layout` enum using `define_enum`
