+++ src/utils.rs
@@ -3,6 +3,35 @@ use std::path::Path;
 
 use serde::Serialize;
 
+#[macro_export]
+macro_rules! define_enum {
+    ($name:ident, $($variant:ident => $str:expr),+ $(,)?) => {
+        #[derive(Serialize, Deserialize, Debug, Clone, clap::ValueEnum)]
+        #[allow(non_camel_case_types)]
+        pub enum $name {
+            $($variant),+
+        }
+
+        impl $name {
+            pub fn to_str(self) -> &'static str {
+                match self {
+                    $($name::$variant => $str),+
+                }
+            }
+        }
+
+        #[allow(dead_code)]
+        impl $name {
+            fn from_str(s: &str) -> std::result::Result<Self, ()> {
+                match s {
+                    $($str => Ok($name::$variant)),+,
+                    _ => Err(()),
+                }
+            }
+        }
+    };
+}
+
 pub fn write_json_to_file<T: Serialize, P: AsRef<Path>>(
     obj: T,
     path: P,
